Update TextLogger class
- Renamed ViolatingAssetLogget class to TextLogger
- Move to new namespace
- Update using references
