+++ Editor/Logging/TextLogger.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3d6bb55ffe03430aac7e4400e995be1b
+timeCreated: 1629251990
